Add Authorization functionality.
Authorization functionality added to mimic a real application. It was a choice
between session cookies and JWTs. Session cookies was deemed the more insecure
choice as it is assumed we are not interacting with a browser so cookies could
be potentially be stolen. (If interacting with a browser, there are ways to add
csrf tokens in forms.) The other choice is Basic Auth but it is not recommended
to use this technique for web based clients.

As a result of addeing Authorization, I've had to make some unit tests be more
unit test-like as they were more resource-like tests before.

Also added a few more integration tests to test the basic api calls.

Added a 'jwtTokenValiditySeconds' in the config.yml to enable configuration of
token expiry